backend/audio_pipeline/requirements.txt
openai-whisper
pyannote.audio>=4.0.1
demucs
shazamio
thefuzz
librosa
soundfile
scipy
scikit-learn
numpy
torch
yt-dlp
